requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
openai
python-dotenv
chromadb

